Update 1.0 -- Can now open accounts and established foundations for most main methods.

+++ src/Bank/Account.java
@@ -7,9 +7,9 @@ public class Account {
     protected static int accountNumber;
     Person accountOwner;
     static Scanner receiver = new Scanner(System.in);
-    static long amount;
+    static double amount;
 
-    public Account(int balance, int accountNumber, Person accountOwner) {
+    public Account(double balance, int accountNumber, Person accountOwner) {
         this.balance = balance;
         this.accountNumber = accountNumber;
         this.accountOwner = accountOwner;
@@ -35,7 +35,7 @@ public class Account {
         Person person = Bank.getPerson();
         if(Bank.verifyAccount(account, person)) {
             System.out.println("Please state the amount of money you wish to deposit (Ex: 16.25).");
-            amount = receiver.nextLong();
+            amount = receiver.nextDouble();
             account.balance = account.balance + amount;
             System.out.println("You have successfully deposited $" + amount + "to account #" + account.accountNumber + ".");
             System.out.println("Your new balance is $" + account.balance + ". Thank you.");
@@ -47,7 +47,7 @@ public class Account {
         Person person = Bank.getPerson();
         if(Bank.verifyAccount(account, person)) {
             System.out.println("Please state the amount of money you wish to withdraw (Ex: 17.62).");
-            amount = receiver.nextLong();
+            amount = receiver.nextDouble();
             account.balance = account.balance - amount;
             System.out.println("You have successfully withdrawn $" + amount + "from account #" + account.accountNumber + ".");
             System.out.println("Your new balance is $" + account.balance + ". Thank you.");
